// Aggregator/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    inputText: ComponentFramework.PropertyTypes.StringProperty;
    Symbol: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    inputText?: string;
}